+++ NEWS.md
@@ -1,3 +1,6 @@
+# CHANGES IN formatR VERSION 1.13
+
+
 # CHANGES IN formatR VERSION 1.12
 
 - Added a new argument `pipe` to `tidy_source()`. If `pipe = TRUE`, it will convert the **magrittr** pipe `%>%` to the base R pipe operator `|>`.